+++ src/onegov.election_day/onegov/election_day/formats/election/wabsti/proporz.py
@@ -195,8 +195,9 @@ def import_file(entities, election, file, mimetype,
     results = {}
 
     # This format has one candiate per entity per line
+    filename = _("Results")
     csv, error = load_csv(
-        file, mimetype, expected_headers=HEADERS
+        file, mimetype, expected_headers=HEADERS, filename=filename
     )
     if error:
         errors.append(error)
@@ -216,7 +217,9 @@ def import_file(entities, election, file, mimetype,
             # Pass the errors and continue to next line
             if line_errors:
                 errors.extend(
-                    FileImportError(error=err, line=line.rownumber)
+                    FileImportError(
+                        error=err, line=line.rownumber, filename=filename
+                    )
                     for err in line_errors
                 )
                 continue
@@ -240,10 +243,12 @@ def import_file(entities, election, file, mimetype,
             candidate.list_id = list.id
 
     # The list connections has one list per line
+    filename = _("List connections")
     if connections_file and connections_mimetype:
         csv, error = load_csv(
             connections_file, connections_mimetype,
-            expected_headers=HEADERS_CONNECTIONS
+            expected_headers=HEADERS_CONNECTIONS,
+            filename=filename
         )
         if error:
             errors.append(error)
@@ -259,7 +264,9 @@ def import_file(entities, election, file, mimetype,
 
                 if line_errors:
                     errors.extend(
-                        FileImportError(error=err, line=line.rownumber)
+                        FileImportError(
+                            error=err, line=line.rownumber, filename=filename
+                        )
                         for err in line_errors
                     )
                     continue
@@ -277,9 +284,11 @@ def import_file(entities, election, file, mimetype,
                         lists[id].connection_id = subconnection.id
 
     # The results file has one elected candidate per line
+    filename = _("Elected Candidates")
     if elected_file and elected_mimetype:
         csv, error = load_csv(
-            elected_file, elected_mimetype, expected_headers=HEADERS_RESULT
+            elected_file, elected_mimetype, expected_headers=HEADERS_RESULT,
+            filename=filename
         )
         if error:
             errors.append(error)
@@ -292,7 +301,8 @@ def import_file(entities, election, file, mimetype,
                     errors.append(
                         FileImportError(
                             error=_("Invalid values"),
-                            line=line.rownumber
+                            line=line.rownumber,
+                            filename=filename
                         )
                     )
                 else:
@@ -305,15 +315,17 @@ def import_file(entities, election, file, mimetype,
                         errors.append(
                             FileImportError(
                                 error=_("Unknown candidate"),
-                                line=line.rownumber
+                                line=line.rownumber,
+                                filename=filename
                             )
                         )
 
     # The stats file has one muncipality per line
+    filename = _("Election statistics")
     if statistics_file and statistics_mimetype:
         csv, error = load_csv(
             statistics_file, statistics_mimetype,
-            expected_headers=HEADERS_STATS
+            expected_headers=HEADERS_STATS, filename=filename
         )
         if error:
             errors.append(error)
@@ -331,7 +343,8 @@ def import_file(entities, election, file, mimetype,
                     errors.append(
                         FileImportError(
                             error=_("Invalid values"),
-                            line=line.rownumber
+                            line=line.rownumber,
+                            filename=filename
                         )
                     )
                 else: