+++ src/onegov.election_day/onegov/election_day/cli.py
@@ -79,7 +79,7 @@ def fetch(group_context):
     return fetch_results
 
 
-@cli.command()
+@cli.command('send-sms')
 @click.argument('username')
 @click.argument('password')
 @click.option('--originator')