+++ src/onegov.election_day/HISTORY.rst
@@ -4,6 +4,9 @@ Changelog
 - Shows the filename of the import errors.
   [msom]
 
+- Shows the filename of the import errors.
+  [msom]
+
 - Hides empty sankey nodes.
   [msom]