+++ src/onegov.election_day/HISTORY.rst
@@ -7,6 +7,9 @@ Changelog
 - Renames the send sms command.
   [msom]
 
+- Adds sentry option for fetch command.
+  [msom]
+
 - Hides empty sankey nodes.
   [msom]